requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# API clients
httpx==0.25.2
cachetools==5.3.2

# Data quality
great-expectations==0.18.8
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
alembic==1.13.1

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Orchestration
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.9.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1

# Logging and monitoring
python-json-logger==2.0.7

# ML (optional extension)
scikit-learn==1.3.2
joblib==1.3.2

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1

